# public/openapi.yaml
openapi: 3.0.3
info:
  title: Seva Sansaar API
  version: "1.0.0"
  description: |
    HTTP API for Seva Sansaar (bookings, businesses, reviews, owner portal, admin).
    Admin routes require header `x-api-key` matching `ADMIN_API_KEY` (and `NEXT_PUBLIC_ADMIN_API_KEY` on the client).
servers:
  - url: /
    description: Same origin as the Next.js app
...
paths:
  /api/bookings:
    post:
      tags: [Bookings]
      summary: Create booking (Firebase Auth OR guest)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                guestBooking: { type: boolean }
                businessId: { type: string }
                businessName: { type: string }
                service: { type: string }
                date: { type: string, description: YYYY-MM-DD }
                timeSlot: { type: string }
                userName: { type: string }
                userPhone: { type: string }
                userEmail: { type: string }
                notes: { type: string }
      responses:
        "201": { description: Created }
        "401": { description: Unauthorized }
        "503": { description: Database unavailable }
...
  /api/bookings/batch:
    post:
      tags: [Bookings]
      summary: Fetch bookings by id list (guest history)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: string }
      responses:
        "200": { description: OK }
  /api/bookings/{id}/status:
    patch:
      tags: [Bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [pending, confirmed, cancelled, completed] }
                guestPhone: { type: string, description: Required for guest cancellations }
      responses:
        "200": { description: OK }
  /api/businesses:
    get:
      tags: [Businesses]
      summary: List businesses
    post:
      tags: [Businesses]
      summary: Create business (admin auth)
      security: [{ bearerAuth: [] }]
  /api/businesses/{id}:
    get:
      tags: [Businesses]
      summary: Get one business
    put:
      tags: [Businesses]
      security: [{ bearerAuth: [] }]
    delete:
      tags: [Businesses]
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT